Be ready to work with archives/RPMs containing data streams
If e.g. a tailoring file or CPE dictionary is used together with a data stream
file the whole content provided to the addon will likely be an archive or an
RPM.

We need to explore the extracted content files a bit more, decide which
content handling class to use and remember that for further decisions.