Exit code 2 from the oscap tool is not an error for us (#1050913)
'man oscap' says:
Normally, the exit status is 0 when operation finished successfully and 1
otherwise. In cases  when oscap  performs  evaluation  of  the system it may
return 2 indicating success of the operation but incompliance of the assessed
system.

Only the success of the operation is important for us in that phase.