Disable GPG checks when installing content to the system
When we are installing the SCAP content as an RPM package we are likely to be
missing some GPG keys. And since there are other means how to check the content
integrity and authenticity plus the fact we have already used the content for the
installation phase, the GPG check can be just turned off to prevent issues like
the one described in the bug.

Resolves: rhbz#1263216