Use the init_done signal instead of wait_all() (#1380224)
Using the wait_all() function provided by the Anaconda thread manager
singleton can result in unexpected behavior, as there might still be
threads unrelated to spoke initialization still running when the
function is called. This might result in unnecessary delay before the
OSCAP addon spoke becomes ready and installation can continue.

So use the init_done signal, which is the proper way to wait for
Anaconda spokes to finish initialization.

Also bump the Anaconda version to one that provides the init_done
signal.

Resolves: rhbz#1380224