Update unit tests after upgrading code to latest Anaconda code
In some of the `rule_handling` methods that handle kickstart rules we
recently started to use new Anaconda API that uses DBUS proxy. See
https://github.com/OpenSCAP/oscap-anaconda-addon/pull/81 However, after
this change our upstream tests started to fail.  The problem with
upstream tests was that they don't replace the DBus proxy by a mock
object and also that they mocked the old classes that aren't used in the
SUT. This PR updates our unit test by mocking the DBus proxy. We keep
the mock object that is returned by DBus.get_proxy, modify its values
and then run the code that uses these values. The test update is
inspired by:
https://github.com/rhinstaller/anaconda/blob/master/tests/nosetests/pyanaconda_tests/partitioning_test.py#L48