Replace invalid patches with a new fixture for the DBus proxies
The fixture proxy_getter patches the method get_proxy of the Anaconda's DBus
object. The fixture creates mocked DBus proxies of the requested DBus objects.
The proxies are kept in a cache and reused when possible. On the first call,
the fixture will call the function set_dbus_defaults that will initialize some
of these mocks.