Replace DataStreamHandler and BenchmarkHandler
The new class SCAPContentHandler can handle SCAP source data
streams, XCCDF files and tailoring files using common code.
The classes DataStreamHandler and BenchmarkHandler are removed.
The purpose of the new class is the same as of the removed
classes. The purpose is to get a list of profiles (their IDs
and descriptions) to be able to show it in the GUI. It is
used only in the GUI.