Refactor and improve readability of get_profiles()
The code that extracts the pointer to root element of an XCCDF
Benchmark has been extracted out to a new private method
`_parse_profiles_from_xccdf`. Also, the conditions in get_profiles()
have been reorganized so that the function now checks for all possible
errors and proceeds only when the checks succeed.