Move the decision of verifying fingerprint
Move the decision whether to verify fingerprint outside from the
`finish_content_fetch` method. The fingerprint will be verified if the
`fingerprint` parameter of the method is set (is not None). This method
is called from 2 places. When calling from _init_data_fetch, the
dest_filename is always not None. And, when calling from
`PrepareValidContent.run()` we can set fingerprint to a non-None value
if the content_dest is set. The reason for this change is that the
`finish_content_fetch` method will not depend on the destination content
file name so it will reduce amount of needed parameters.